api-service/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
fastapi==0.105.0
fastapi-cli==0.0.7
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.4
packaging==24.2
preshed==3.0.9
PyAudio==0.2.14
pycparser==2.22
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
scipy==1.15.2
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speech-recognition-api==0.1.1
SpeechRecognition==3.14.1
srsly==2.5.1
starlette==0.27.0
thinc==8.3.4
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.25.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
